Remove all @will-shaw stuff. Sorry bud.
I no longer use any of it, some of it was quite specific (eg the boot
time thing), it was adding a lot to complexity.

Fonts are easy to re-add. VSCode stuff is still there, for when neovim
integration gets good enough for me to start using it again.

The method of dotfile management being used for the Shaw stuff (copying to a custom dir then sourcing from there) was awkward,
and is better suited to just using stow or fresh or one of the other myriad
methods other people use.